Prevent errors when faceting over an attribute in a foreign table
The faceting query should always return counts of fields in the current
table.  There were cases when a faceting operation was throwing an error
due to invalid join condition in the main query.  The correct behavior
is to return an empty dictionary.  For example, faceting on the 'study`
table over the attribute `biosample_id` returns no results because no
studies contain that field.