Add the ability to query by the existence of multiple omics types
This adds a new denormalized column called `multiomics` to the biosample
table.  This is an integer representing a bit mask of omics processing
types available for the sample.  Bit values are encoded in an enum:

```python
class MultiomicsValue(Enum):
    mb = 0b10000
    mg = 0b01000
    mp = 0b00100
    mt = 0b00010
    om = 0b00001
```

A new schema for querying this information is also implemented:

```python
{
  "table": "biosample",
  "field": "multiomics",
  "op": "has",
  "value": 6
}
```
This will query for biosamples that have **both** metatranscriptome and
metaproteomic omics_processing.  (They might have others as well.)