Use `term.id` for MIXS triad objects
This is a change in schema 7.0.0. Some values of these slots
(e.g. `env_medium`) would be a string, followed by the ID in
brackets. Now there is a property of the slot called `term`.
`term` is an object with an `id` and a `name`. We can use the
`id` to look up the envo term.