Update connection from sample to study
NMDC schema changed the slot that links a Biosample to a study.
Previously the relationship was contained in the `part_of` slot on class
`Biosample`. That slot has been renamed to `associated_studies`. Also,
temporarily disable the backup link through omics_processing, as that
relationship has changed in a more complex way.