Update ingest for data_generation objects
Formerly known as omics_procecssing records, there were some small
schema tweaks that needed to be reflected in ingest. Note how the
process of obtaining input biosamples is simpler since we now only need
to query one collection of processes. This is a result of the change in
database structure that puts related objects into the same collection.