Add new workflow models for MetaT
This essentially duplicates existing functionality for analogous
Metagenome workflows. That is, MetagenomeAnnotation and
MetagenomeAssembly now have Metatranscriptome counterparts. Also
includes a migration to create the new tables. This does not attempt to
modify query logic for MetatranscriptomeAnnotations (i.e. genome
function searches won't connect to that table). That could be future
work.

This is in support of some changes to the NMDC schema that fully realize
these classes as collections.